chore: extract ToolMetadata to metadata module
Move ToolMetadata, EvidenceType and SideEffectLevel out of BaseTool into core/tool_framework/metadata.py and import them where needed. Updated core.tool_framework.base, registered_tool, tool_decorator, and tests to use the new metadata module. Keeps the validation/schema separate from the abstract BaseTool so the metadata contract can evolve independently; no behavioral changes.
